Add HttpVersion parameter to Invoke-WebRequest
fixes by review
- remove HttpVersionUtils
- refactor pester test
